fix: 리뷰 API 세션 로그인 체크 추가

+++ src/main/java/com/sparta/backoffice/review/controller/ReviewController.java
@@ -15,6 +15,7 @@ import com.sparta.backoffice.review.dto.ReviewGetAllResponse;
 import com.sparta.backoffice.review.dto.ReviewGetOneResponse;
 import com.sparta.backoffice.review.service.ReviewService;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -26,25 +27,39 @@ public class ReviewController {
 	// 리뷰 목록 페이징 및 필터 조회
 	@GetMapping
 	public ResponseEntity<Page<ReviewGetAllResponse>> getReviews(
-		@ModelAttribute ReviewGetAllRequest request
+		@ModelAttribute ReviewGetAllRequest request,
+		HttpSession session
 	) {
+		Long adminId = (Long)session.getAttribute("adminId");
+		if (adminId == null) {
+			throw new IllegalArgumentException("로그인이 필요합니다.");
+		}
 		return ResponseEntity.status(HttpStatus.OK).body(reviewService.getReviews(request));
 	}
 
 	// 리뷰 상세 조회
 	@GetMapping("/{id}")
 	public ResponseEntity<ReviewGetOneResponse> getReview(
-		@PathVariable Long id
+		@PathVariable Long id,
+		HttpSession session
 	) {
+		Long adminId = (Long)session.getAttribute("adminId");
+		if (adminId == null) {
+			throw new IllegalArgumentException("로그인이 필요합니다.");
+		}
 		return ResponseEntity.status(HttpStatus.OK).body(reviewService.getReview(id));
 	}
 
 	// 리뷰 삭제
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteReview(
-		@PathVariable Long id) {
+		@PathVariable Long id,
+		HttpSession session) {
+		Long adminId = (Long)session.getAttribute("adminId");
+		if (adminId == null) {
+			throw new IllegalArgumentException("로그인이 필요합니다.");
+		}
 		reviewService.deleteReview(id);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 	}
-
 }
